Code cleaning and fix small mistakes
