+++ packages/replica-server/src/config/queue.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { BullModule } from '@nestjs/bullmq';
 import { ConfigModule, ConfigService } from '@nestjs/config';
 import { getRedisConfig } from 'src/config/datasource';
-import { QueueName } from './const';
+import { QueueName } from '../core/bus/const';
 
 @Module({
   imports: [
