+++ packages/replica-server/src/core/bus/bus.module.ts
@@ -1,16 +1,11 @@
 import { Module } from '@nestjs/common';
-import {
-  CommonEventProcessor,
-  CommonUnsyncedEventProcessor,
-} from './common.processor';
 import { TransferModule } from '../event-transfer/transfer.module';
-import { QueueModule } from './queue.module';
+import { QueueModule } from '../../config/queue.module';
+import { SyncQueueProcessor } from './processors/sync-queue.processor';
+import { UnsyncQueueProcessor } from './processors/unsynced-queue.processor';
 
 @Module({
-  imports: [
-    QueueModule,
-    TransferModule, // Import contract modules
-  ],
-  providers: [CommonEventProcessor, CommonUnsyncedEventProcessor],
+  imports: [QueueModule, TransferModule],
+  providers: [SyncQueueProcessor, UnsyncQueueProcessor],
 })
 export class BusModule {}
