+++ packages/replica-server/src/core/marketplace/consts.ts
@@ -1,8 +1,3 @@
 export enum MarketplaceJobName {
   NftListed = 'nftListed',
 }
-
-export enum MarketplaceQueueName {
-  marketplaceEvents = 'marketplace_events',
-  unsyncedMarketplaceEvents = 'unsynced_marketplace_events',
-}
