+++ packages/replica-server/src/core/marketplace/handlers/nft-listed-event.handler.ts
@@ -1,5 +1,5 @@
 import { Injectable } from '@nestjs/common';
-import { MarketplaceEventService } from '../marketplace-event.service';
+import { MarketplaceEventService } from '../services/marketplace-event.service';
 import { ContractEventHandler } from 'src/core/bus/types';
 
 @Injectable()
@@ -7,7 +7,7 @@ export class NftListedHandler implements ContractEventHandler {
   constructor(private marketplaceEventService: MarketplaceEventService) {}
 
   async handle(data: any, isUnsyncedEvent: boolean = false): Promise<void> {
-    console.log('Handling nft listed event:', data);
+    console.log('Handling nft listed event:', data, isUnsyncedEvent);
     await this.marketplaceEventService.saveListing(data, isUnsyncedEvent);
   }
 }
