+++ packages/replica-server/src/core/marketplace/marketplace.module.ts
@@ -1,24 +1,17 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { MarketplaceEventService } from './marketplace-event.service';
-// import {
-//   MarketplaceEventConsumer,
-//   UnsyncedMarketplaceEventConsumer,
-// } from './processors/marketplace-event.processor';
+import { MarketplaceEventService } from './services/marketplace-event.service';
 import { Transaction } from 'ethers';
-import { BullModule } from '@nestjs/bullmq';
-import { MarketplaceQueueName } from './consts';
-import { ConfigModule, ConfigService } from '@nestjs/config';
-import { getRedisConfig } from 'src/config/datasource';
-import { MarketplaceSyncService } from './marketplace-sync.service';
-import { MarketplaceContractService } from './marketplace-contract.service';
-import { MarketplacePublisherService } from './marketplace-publisher.service';
+import { ConfigModule } from '@nestjs/config';
+import { MarketplaceSyncService } from './services/marketplace-sync.service';
+import { MarketplaceContractService } from './services/marketplace-contract.service';
+import { MarketplacePublisherService } from './services/marketplace-publisher.service';
 import { RedisModule } from 'src/config/redis.module';
 import { Listing } from './entities/listing.entity';
-import { MarketplaceListnerService } from './marketplace-listner.service';
+import { MarketplaceListnerService } from './services/marketplace-listner.service';
 import { BlockchainModule } from '../blockchain/blockchain.module';
 import { NftListedHandler } from './handlers/nft-listed-event.handler';
-import { QueueModule } from '../bus/queue.module';
+import { QueueModule } from '../../config/queue.module';
 
 @Module({
   imports: [
@@ -34,9 +27,7 @@ import { QueueModule } from '../bus/queue.module';
     MarketplacePublisherService,
     MarketplaceSyncService,
     MarketplaceEventService,
-    // MarketplaceEventConsumer,
     MarketplaceListnerService,
-    // UnsyncedMarketplaceEventConsumer,
   ],
   exports: [TypeOrmModule, NftListedHandler],
 })
