+++ packages/replica-server/src/core/marketplace/services/marketplace-contract.service.ts
@@ -4,7 +4,7 @@ import {
   openMarketplaceContractAbi,
   OpenMarketplace,
 } from '@nft-open-marketplace/interface';
-import { BlockchainContractsService } from '../blockchain/blockchain-contracts.service';
+import { BlockchainContractsService } from '../../blockchain/blockchain-contracts.service';
 
 @Injectable()
 export class MarketplaceContractService {
