+++ packages/replica-server/src/core/marketplace/services/marketplace-sync.service.ts
@@ -5,8 +5,8 @@ import {
   HandleContractLog,
   GetContractInterface,
   ReplicateService,
-} from '../sync/sync.decorators';
-import { Replicable } from '../sync/sync.types';
+} from '../../sync/sync.decorators';
+import { Replicable } from '../../sync/sync.types';
 import { MarketplaceContractService } from './marketplace-contract.service';
 import { MarketplacePublisherService } from './marketplace-publisher.service';
 import { NftListedEvent } from '@nft-open-marketplace/interface/dist/esm/typechain-types/contracts/OpenMarketplace.sol/OpenMarketplace';
@@ -43,7 +43,6 @@ export class MarketplaceSyncService implements Replicable {
     log: ethers.Log,
     logDescription: LogDescription,
   ) {
-    console.log('Market log', logDescription.args);
     if (logDescription.name === 'NftListed') {
       const args = (logDescription as unknown as NftListedEvent.Log).args;
       console.log('Publish NftListed unsync. Args', args);
