+++ packages/replica-server/src/core/nft/consts.ts
@@ -1,8 +1,3 @@
 export enum NftEventJobName {
   Transfer = 'transfer',
 }
-
-// export enum NftQueueName {
-//   nftEvents = 'nft_events',
-//   unsyncedNftEvents = 'unsynced_nft_events',
-// }
