+++ packages/replica-server/src/core/nft/services/nft-contract.service.ts
@@ -4,7 +4,7 @@ import {
   openMarketplaceNFTContractAbi,
   OpenMarketplaceNFT,
 } from '@nft-open-marketplace/interface';
-import { BlockchainContractsService } from '../blockchain/blockchain-contracts.service';
+import { BlockchainContractsService } from '../../blockchain/blockchain-contracts.service';
 
 @Injectable()
 export class NftContractService {
