+++ packages/replica-server/src/core/nft/services/nft-listner.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from '@nestjs/common';
 import { ContractEventPayload } from 'ethers';
 import { NftPublisherService } from './nft-publisher.service';
-import { BlockchainTransportService } from '../blockchain/blockchain-transport.service';
+import { BlockchainTransportService } from '../../blockchain/blockchain-transport.service';
 import { NftContractService } from './nft-contract.service';
 import { OpenMarketplaceNFT } from '@nft-open-marketplace/interface';
 
@@ -22,7 +22,6 @@ export class NftListnerService {
   }
 
   private async listenNode() {
-    console.log('LISTEN NODE: TRANSFER');
     this.contract.on(
       'Transfer' as any,
       async (
@@ -31,8 +30,6 @@ export class NftListnerService {
         tokenId: BigInt,
         eventPayload: ContractEventPayload,
       ) => {
-        console.log('transfer event -->', from, to, tokenId);
-
         await this.publisherService.publishTransferEventData({
           from,
           to,
